+++ packages/leaders/browser/.eslintrc.js
