+++ packages/leaders/browser/company-social-link.vue
