+++ packages/leaders/components/all.marko
@@ -7,19 +7,17 @@ $ const enabled = defaultValue(site.get("leaders.enabled"), false);
   <marko-web-leaders props={
     sectionAlias: site.get("leaders.alias"),
     open: null,
-    expanded: false,
-    headerImgSrc: site.get("leaders.header.imgSrc"),
-    headerImgAlt: site.get("leaders.title"),
-    calloutValue: site.get("leaders.calloutValue"),
-    columns: 1,
+    expanded: true,
+    contextual: true,
+    displayHeader: false,
+    columns: 3,
     offsetTop: site.get("leaders.offsetTop", 105),
     promotionLimit: site.get("leaders.promotionLimit", 3),
     videoLimit: site.get("leaders.videoLimit", 2),
     viewAllHref: '/leaders',
     mediaQueries: [
-      { prop: "open", value: "right", query: "(min-width: 1250px)" },
-      { prop: "columns", value: 2, query: "(min-width: 700px)" },
-      { prop: "displayCallout", value: false, query: "(max-width: 660px)" },
+      { prop: "columns", value: 2, query: "(max-width: 991px)" },
+      { prop: "columns", value: 1, query: "(max-width: 700px)" },
     ],
-  } />
+  }/>
 </if>
