+++ packages/leaders/components/company-social-link.marko
