+++ packages/leaders/components/contextual.marko
@@ -1,23 +1,24 @@
-import defaultValue from "@base-cms/marko-core/utils/default-value";
-
 $ const { site } = out.global;
-$ const enabled = defaultValue(site.get("leaders.enabled"), false);
+$ const { contentId } = input;
 
-<if(enabled)>
+<if(contentId)>
   <marko-web-leaders props={
+    contentId,
     sectionAlias: site.get("leaders.alias"),
-    open: null,
-    expanded: true,
-    contextual: true,
-    displayHeader: false,
-    columns: 3,
+    open: "left",
+    headerImgSrc: site.get("leaders.header.imgSrc"),
+    headerImgAlt: site.get("leaders.title"),
+    displayCallout: false,
     offsetTop: site.get("leaders.offsetTop", 105),
     promotionLimit: site.get("leaders.promotionLimit", 3),
     videoLimit: site.get("leaders.videoLimit", 2),
     viewAllHref: '/leaders',
     mediaQueries: [
+      { prop: "displayCallout", value: true, query: "(min-width: 701px)" },
+      { prop: "displayCallout", value: false, query: "(min-width: 992px)" },
       { prop: "columns", value: 2, query: "(max-width: 991px)" },
+      { prop: "open", value: null, query: "(max-width: 991px)" },
       { prop: "columns", value: 1, query: "(max-width: 700px)" },
     ],
-  } />
+  }/>
 </if>
