+++ packages/leaders/components/gtm-tracker.marko
@@ -1,0 +1,1 @@
+<marko-web-browser-component name="CommonGTMTrackLeaders" />
