+++ packages/leaders/components/home-page.marko
@@ -1,27 +1,25 @@
 import defaultValue from "@base-cms/marko-core/utils/default-value";
 
 $ const { site } = out.global;
-$ const { contentId } = input;
 $ const enabled = defaultValue(site.get("leaders.enabled"), false);
 
-<if(enabled && contentId)>
+<if(enabled)>
   <marko-web-leaders props={
-    contentId,
     sectionAlias: site.get("leaders.alias"),
-    open: "left",
+    open: null,
+    expanded: false,
     headerImgSrc: site.get("leaders.header.imgSrc"),
     headerImgAlt: site.get("leaders.title"),
-    displayCallout: false,
+    calloutValue: site.get("leaders.calloutValue"),
+    columns: 1,
     offsetTop: site.get("leaders.offsetTop", 105),
     promotionLimit: site.get("leaders.promotionLimit", 3),
     videoLimit: site.get("leaders.videoLimit", 2),
     viewAllHref: '/leaders',
     mediaQueries: [
-      { prop: "displayCallout", value: true, query: "(min-width: 701px)" },
-      { prop: "displayCallout", value: false, query: "(min-width: 992px)" },
-      { prop: "columns", value: 2, query: "(max-width: 991px)" },
-      { prop: "open", value: null, query: "(max-width: 991px)" },
-      { prop: "columns", value: 1, query: "(max-width: 700px)" },
+      { prop: "open", value: "right", query: "(min-width: 1250px)" },
+      { prop: "columns", value: 2, query: "(min-width: 700px)" },
+      { prop: "displayCallout", value: false, query: "(max-width: 660px)" },
     ],
-  } />
+  }/>
 </if>
