+++ packages/leaders/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@ac-business-media/package-leaders",
+  "version": "1.16.2",
+  "author": "Josh Worden <josh@limit0.io>",
+  "repository": "https://github.com/base-cms-websites/ac-business-media/tree/master/packages/leaders",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "lint": "eslint --ext .js --ext .vue --max-warnings 5 --ignore-path ../../.eslintignore ./",
+    "test": "yarn lint"
+  },
+  "dependencies": {
+    "@base-cms/marko-core": "^1.25.5",
+    "@base-cms/marko-web": "^1.25.5",
+    "@base-cms/marko-web-icons": "^1.24.0",
+    "@base-cms/marko-web-leaders": "^1.25.7",
+    "@base-cms/object-path": "^1.25.0",
+    "graphql-tag": "^2.10.1",
+    "object-path": "^0.11.4"
+  },
+  "devDependencies": {
+    "babel-eslint": "^10.0.3"
+  }
+}
