+++ packages/leaders/scss/index-page.scss
@@ -1,0 +1,35 @@
+.leaders-index-page {
+  &__header-wrapper {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    width: 100%;
+  }
+
+  &__title {
+    margin-bottom: .75rem;
+    font-size: 34px;
+  }
+
+  &__header {
+    margin-bottom: .75rem;
+  }
+
+  &__logo {
+    display: flex;
+    padding: .75rem;
+    margin-right: .75rem;
+    margin-bottom: .75rem;
+    background-color: $leaders-primary-color;
+  }
+
+  &__body,
+  &__description {
+    > p {
+      margin-bottom: .25rem;
+      &:last-child {
+        margin-bottom: 0;
+      }
+    }
+  }
+}
