+++ packages/leaders/scss/index.scss
@@ -1,0 +1,3 @@
+@import url("https://fonts.googleapis.com/css?family=IBM+Plex+Sans:300,400,500,600,700&display=swap");
+@import "../../node_modules/@base-cms/marko-web-leaders/scss/leaders";
+@import "./index-page.scss"
