+++ packages/leaders/templates/leaders.marko
@@ -1,0 +1,51 @@
+import hierarchyAliases from "@base-cms/marko-web/utils/hierarchy-aliases";
+import GAM from "../../../config/gam";
+
+$ const { site, config } = out.global;
+$ const { id, alias, name, pageNode } = data;
+
+<marko-web-website-section-page-layout id=id alias=alias name=name>
+  <@head>
+    <marko-web-gtm-website-section-context|{ context }| alias=alias>
+      <marko-web-gtm-push data=context />
+    </marko-web-gtm-website-section-context>
+    <marko-web-resolve-page|{ data: section }| node=pageNode>
+      $ const aliases = hierarchyAliases(section);
+      $ const adSlots = {
+          "gpt-ad-leaderboard": GAM.getAdUnit({ name: "leaderboard", aliases }),
+        }
+       <marko-web-gam-slots slots=adSlots />
+    </marko-web-resolve-page>
+  </@head>
+  <@above-container>
+    <marko-web-gam-display-ad id="gpt-ad-leaderboard" modifiers=["above-container"] />
+    <marko-web-gtm-slot prefix=site.get("gtm.slotPrefix") name="home-top-below-head" modifiers=["above-container"] />
+  </@above-container>
+  <@page>
+    <marko-web-page-wrapper>
+      <@section>
+        <div class="leaders-index-page">
+          <div class="leaders-index-page__header-wrapper">
+            <div class="leaders-index-page__logo">
+              <marko-web-img
+                src=site.get("leaders.header.imgSrc")
+                alt=site.get("leaders.title")
+              />
+            </div>
+            <div class="leaders-index-page__header">
+              <h1 class="leaders-index-page__title">
+                ${site.get("leaders.title")}
+              </h1>
+              <div class="leaders-index-page__description">
+                <p>Welcome to <em>${config.siteName()}'s</em> ${site.get("leaders.title")} program.</p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </@section>
+      <@section>
+        <leaders-all />
+      </@section>
+    </marko-web-page-wrapper>
+  </@page>
+</marko-web-website-section-page-layout>
