+++ packages/refresh-theme/browser/index.js
@@ -3,27 +3,14 @@ import GTM from '@base-cms/marko-web-gtm/browser';
 import GAM from '@base-cms/marko-web-gam/browser';
 import GCSE from '@base-cms/marko-web-gcse/browser';
 import Inquiry from '@base-cms/marko-web-inquiry/browser';
-import Leaders from '@base-cms/marko-web-leaders/browser';
+import Leaders from '@ac-business-media/package-leaders/browser';
 import Common from '@ac-business-media/package-common/browser';
 import SocialSharing from '@base-cms/marko-web-social-sharing/browser';
 import PhotoSwipe from '@base-cms/marko-web-photoswipe/browser';
 import RevealAd from '@base-cms/marko-web-reveal-ad/browser';
 import Radix from '@base-cms/marko-web-radix/browser';
 
-const GTMTrackLeaders = () => import(/* webpackChunkName: "gtm-track-leaders" */ './gtm-track-leaders.vue');
-const LeadersCompanyWebsiteLink = () => import(/* webpackChunkName: "leaders-company-website-link" */ './company-website-link.vue');
-const LeadersCompanySocialLink = () => import(/* webpackChunkName: "leaders-company-social-link" */ './company-social-link.vue');
-
 export default (Browser) => {
-  const { EventBus } = Browser;
-  Browser.register('GTMTrackLeaders', GTMTrackLeaders, { provide: { EventBus } });
-  Browser.register('LeadersCompanyWebsiteLink', LeadersCompanyWebsiteLink, {
-    on: { action: (...args) => EventBus.$emit('leaders-action', ...args) },
-  });
-  Browser.register('LeadersCompanySocialLink', LeadersCompanySocialLink, {
-    on: { action: (...args) => EventBus.$emit('leaders-action', ...args) },
-  });
-
   DefaultTheme(Browser);
   Leaders(Browser);
   GTM(Browser);
