+++ packages/refresh-theme/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@ac-business-media/package-common": "^1.12.0",
     "@ac-business-media/package-directory": "^1.12.0",
+    "@ac-business-media/package-leaders": "^1.16.2",
     "@base-cms/marko-core": "^1.25.5",
     "@base-cms/marko-web": "^1.25.5",
     "@base-cms/marko-web-gam": "^1.25.10",
@@ -19,7 +20,6 @@
     "@base-cms/marko-web-gtm": "^1.25.0",
     "@base-cms/marko-web-icons": "^1.24.0",
     "@base-cms/marko-web-inquiry": "^1.26.0",
-    "@base-cms/marko-web-leaders": "^1.25.7",
     "@base-cms/marko-web-native-x": "^1.25.0",
     "@base-cms/marko-web-photoswipe": "^1.25.0",
     "@base-cms/marko-web-radix": "^1.26.0",
