+++ packages/refresh-theme/templates/content/index.marko
@@ -225,7 +225,7 @@ $ const injectAds = ["article", "news", "press-release", "blog"].includes(type) 
                 <@native-x index=0 name="default" aliases=aliases />
               </refresh-theme-latest-in-section-block>
 
-              <refresh-theme-leaders-contextual content-id=id />
+              <leaders-contextual content-id=id />
 
               <if(!isShortForm)>
                 <marko-web-gam-display-ad
