+++ packages/refresh-theme/templates/website-section/leaders.marko
@@ -45,7 +45,7 @@ $ const {
           </div>
         </@section>
         <@section>
-          <refresh-theme-leaders-all />
+          <leaders-all />
         </@section>
       </marko-web-page-wrapper>
     </marko-web-resolve-page>
